requirements.txt
python-dotenv
google-generativeai
openai-whisper
sounddevice
scipy
numpy<2.0
pynput